refactor(agent): remove BusTransport and session dependency
