+++ packages/agent/src/runtime/messenger/index.ts
@@ -1,6 +1,5 @@
 export { AgentMessenger } from "./messenger";
-export { BusTransport } from "./transport";
-export type { Transport } from "./transport";
+export type { Transport } from "./messenger";
 export type { AgentMessengerOptions, RequestOptions } from "./messenger";
 export { InstanceRegistry } from "./instance-registry";
 export type { AgentInstance, InstanceStatus } from "./instance-registry";
